requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-admin-interface==0.17.3
django-ckeditor==6.0.0
django-colorfield==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-storages==1.11.1
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-decouple==3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
whitenoise==5.3.0
